refactor: ShowAdminUseCase 생성
