+++ app/domain/show-domain/src/main/java/org/example/usecase/artist/ArtistUseCase.java
@@ -151,7 +151,5 @@ public class ArtistUseCase {
         return artistRepository.findById(id)
             .orElseThrow(NoSuchElementException::new);
     }
-
-
 }
